Extract each `tx new` command's integration tests to a dedicated file.
