+++ cmd/crates/soroban-test/tests/it/integration/tx/general.rs
@@ -6,10 +6,6 @@ use crate::integration::util::{
     deploy_contract, test_address, DeployKind, DeployOptions, HELLO_WORLD,
 };
 
-pub mod decode_encode;
-pub mod fetch;
-pub mod operations;
-
 #[tokio::test]
 async fn simulate() {
     let sandbox = &TestEnv::new();
