+++ cmd/crates/soroban-test/tests/it/integration/tx/mod.rs
@@ -1,0 +1,15 @@
+pub mod account_merge;
+pub mod bump_sequence;
+pub mod change_trust;
+pub mod claimable_balance;
+pub mod clawback;
+pub mod create_account;
+pub mod decode_encode;
+pub mod fetch;
+pub mod general;
+pub mod manage_data;
+pub mod offers;
+pub mod path_payments;
+pub mod payment;
+pub mod set_options;
+pub mod sponsorship;
